refactor: code quality improvements and component extraction
- Parallelize database queries in visit page for better performance
- Fix typo: addSurgicaActsForm → addSurgicalActsForm
- Extract print templates into separate components (PrintBillTemplate,
  PrintPrescriptionTemplate, PrintReportTemplate, PrintCertificateTemplate)
- Extract table row components (InventorySaleRow, AnimalTableRow)
- Add ARIA labels and error messages to TextField and SelectField
- Create form action wrapper utility for standardized error handling
- Remove unused 'remove' dependency
- Update validation workflow in documentation
