+++ package.json
@@ -57,7 +57,6 @@
 		"moment": "^2.30.1",
 		"photoswipe": "^5.4.4",
 		"pocketbase": "^0.21.3",
-		"remove": "^0.1.5",
 		"svelte-printpdf": "^1.0.2-metadata",
 		"svelte-select": "^5.8.3",
 		"svelte-sonner": "^0.3.26",
